+++ server/project-service/src/main/java/edu/illinois/ncsa/incore/service/project/models/HazardResource.java
@@ -47,4 +47,13 @@ public class HazardResource extends ProjectResource {
     public Date getDate() {
         return date;
     }
+
+    public boolean matchesSearchText(String text) {
+        String lowerCaseText = text.toLowerCase();
+        return (this.name != null && this.name.toLowerCase().contains(lowerCaseText)) ||
+            (this.description != null && this.description.toLowerCase().contains(lowerCaseText)) ||
+            (this.creator != null && this.creator.toLowerCase().contains(lowerCaseText)) ||
+            (this.owner != null && this.owner.toLowerCase().contains(lowerCaseText)) ||
+            (this.type != null && this.getType().toString().toLowerCase().contains(lowerCaseText));
+    }
 }