+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - Implement Project endpoints [#307](https://github.com/IN-CORE/incore-services/issues/307)
 - Add pagination and filters to datasets, dfr3mappings, hazards, workflows, and visualizations endpoint belong to a project [#318](https://github.com/IN-CORE/incore-services/issues/318)
+- Search endpoints for resources belong to a project [#320](https://github.com/IN-CORE/incore-services/issues/320)
 
 ### Changed
 - Geotools to version 31.3 [#249](https://github.com/IN-CORE/incore-services/issues/249)